Reorganise project struct
