+++ configs/config_deepvoxels_greek.txt
