+++ configs/config_lego.txt
