+++ configs/config_lego_rotation.txt
