+++ configs/config_lego_rotation_fine.txt
