--- download_example_weights.sh
@@ -1,6 +1,0 @@
-mkdir -p logs
-cd logs
-wget https://people.eecs.berkeley.edu/~bmild/nerf/fern_example_weights.zip
-unzip fern_example_weights.zip
-wget https://people.eecs.berkeley.edu/~bmild/nerf/lego_example_weights.zip
-unzip lego_example_weights.zip
