+++ run_nerf_r.py
@@ -13,9 +13,9 @@ import json
 import random
 import time
 from run_nerf_helpers import *
-from load_llff import load_llff_data
-from load_deepvoxels import load_dv_data
-from load_blender import load_blender_data
+from utils.load_llff import load_llff_data
+from utils.load_deepvoxels import load_dv_data
+from utils.load_blender import load_blender_data
 tf.compat.v1.enable_eager_execution()
